refactor(sandbox): extract shared ThemePalettePreview component
- Create reusable ThemePalettePreview for theme palette pages
- Refactor Y2K palette page from 809 → 184 lines using shared component
- Fix providers.tsx to import y2kTheme from /built (matches other themes)
- Standard sections: Tonal, Core, TextRamp, Badges, Banners, Inputs, Buttons, Surfaces
- Theme-specific extras passed via extraSections prop
