+++ apps/sandbox/src/app/providers.tsx
@@ -11,7 +11,7 @@ import {dailyTheme} from '@xds/theme-daily/built';
 import {stoneTheme} from '@xds/theme-stone/built';
 import {gothicTheme} from '@xds/theme-gothic/built';
 import {chocolateTheme} from '@xds/theme-chocolate/built';
-import {y2kTheme} from '@xds/theme-y2k';
+import {y2kTheme} from '@xds/theme-y2k/built';
 import type {XDSDefinedTheme, ThemeMode} from '@xds/core/theme';
 
 const themes: Record<string, XDSDefinedTheme> = {
